gemini/multimodal-live-api/native-audio-websocket-demo-apps/react-demo-app/requirements.txt
websockets>=12.0
google-auth>=2.23.0
certifi>=2023.7.22
requests>=2.31.0
python-dotenv>=1.0.0